Try moving the filesystem tests into their own workflow given the extra deps
