+++ tox.ini
@@ -16,7 +16,16 @@ deps =
 description = adapter plugin functional testing
 skip_install = True
 passenv = *
-commands = {envpython} -m pytest {posargs} tests/functional
+commands = {envpython} -m pytest {posargs} tests/functional/adapter
+deps =
+  -rdev-requirements.txt
+  -e.
+
+[testenv:{fsspec,py37,py38,py39,py310,py311,py}]
+description = adapter plugin functional testing
+skip_install = True
+passenv = *
+commands = {envpython} -m pytest {posargs} tests/functional/fsspec
 deps =
   -rdev-requirements.txt
   -e.
